backend/requirements.txt
flask>=3.1.1,<4.0.0
flask-smorest>=0.46.1,<0.47.0
python-dotenv>=0.9.0,<1.0.0
google-genai>=1.27.0,<2.0.0
gunicorn>=23.0.0,<24.0.0
flask-cors>=6.0.1,<7.0.0
